File I/O

PowerWaterModels.parse_files — Function

parse_files(power_path, water_path, link_path)

Parses power, water, and linking data from power_path, water_path, and link_path, respectively, into a single data dictionary. Returns a PowerWaterModels multi-infrastructure data structure keyed by the infrastructure type it.

PowerWaterModels.parse_json — Function

parse_json(path)

Parses a JavaScript Object Notation (JSON) file from the file path path and returns a PowerWaterModels data structure that links power and water networks (a dictionary of data).